// src/domain/enums/account.ts
export enum AccountType {
  USER_WALLET = 'USER_WALLET',
  MERCHANT_REVENUE = 'MERCHANT_REVENUE',
  SYSTEM_SUSPENSE = 'SYSTEM_SUSPENSE',
  SYSTEM_FEES = 'SYSTEM_FEES',
  SYSTEM_EXTERNAL_SOURCE = 'SYSTEM_EXTERNAL_SOURCE',
  SYSTEM_EXTERNAL_SINK = 'SYSTEM_EXTERNAL_SINK',
}

export enum AccountStatus {
  ACTIVE = 'ACTIVE',
  FROZEN = 'FROZEN',
  CLOSED = 'CLOSED',
}

export enum AccountOwnerType {
  USER = 'USER',
  ORGANIZATION = 'ORGANIZATION',
  SYSTEM = 'SYSTEM',
}

export enum AccountNormalBalance {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}